Updated RESNET18 & PyTorch-ResNet18

+++ Tutorials/RESNET18/files/scripts/clean_all.sh
@@ -5,7 +5,7 @@
 
 ## Author: Daniele Bagni, AMD/Xilinx Inc
 
-## date 24 May 2023
+## date 10 Aug 2023
 
 
 # clean folders
@@ -21,15 +21,45 @@ for file in $(find . -name "*.*~"); do
     echo ${file}
     rm -f ${file}
 done
-rm -r ./build/compiled_v70
-rm -r ./build/compiled_vck190
-rm -r ./build/compiled_zcu102
-rm -r ./build/compiled_vek280
-rm -r ./build/compiled_vck5000
-rm -r ./build/log
-rm -r ./build/quantized
-rm -r ./build/target*
-#rm -r ./build/dataset
+if [ ! -d "./build" ]; then
+    mkdir -p ./build
+    mkdir -p ./build/float
+    mkdir -p ./build/dataset
+    mkdir -p ./build/quantized
+    mkdir -p ./build/log
+    mkdir -p ./build/compiled_v70
+    mkdir -p ./build/compiled_vck190
+    mkdir -p ./build/compiled_zcu102
+    mkdir -p ./build/compiled_vek280
+    mkdir -p ./build/compiled_vck5000    
+else
+    rm -r ./build/compiled_v70
+    rm -r ./build/compiled_vck190
+    rm -r ./build/compiled_zcu102
+    rm -r ./build/compiled_vek280
+    rm -r ./build/compiled_vck5000
+    rm -r ./build/log
+    rm -r ./build/quantized
+    rm -r ./build/target*
+    ### if you want to rebuild the CIFAR10 dataset uncomment next line
+    rm -r ./build/dataset
+    rm -r ./build/float    
+    mkdir -p ./build/float
+    mkdir -p ./build/dataset
+    mkdir -p ./build/quantized
+    mkdir -p ./build/log
+    mkdir -p ./build/compiled_v70
+    mkdir -p ./build/compiled_vck190
+    mkdir -p ./build/compiled_zcu102
+    mkdir -p ./build/compiled_vek280
+    mkdir -p ./build/compiled_vck5000    
+fi
+rm -rf ./modelzoo/*resnet50*
+rm -rf ./modelzoo/ImageNet/val_dataset
+#rm -rf ./modelzoo/ImageNet/ILSVRC2012_img_val.tar
+rm -f  ./modelzoo/ImageNet/val_dataset.zip
+rm -f  ./target/imagenet/val_dataset.zip
+
 # clean target folder
 cd ./target/cifar10
 rm -f *.xmodel
@@ -43,21 +73,9 @@ rm -f ./resnet*imagenet
 rm -f ./get_dpu_fps
 rm -rf val_dataset
 cd ../../
+echo " CUR DUR " ${PWD}
 
 
-#rm -r ./build/float # if you remove it also the trained models will be deleted!
-#mkdir -p ./build/float
-#mkdir -p ./build/dataset
-
-mkdir -p ./build/quantized
-mkdir -p ./build/log
-mkdir -p ./build/compiled_v70
-mkdir -p ./build/compiled_vck190
-mkdir -p ./build/compiled_zcu102
-mkdir -p ./build/compiled_vek280
-mkdir -p ./build/compiled_vck5000
-
-: '
 # dos2unix to all txt files
 echo " "
 echo "----------------------------------------------------------------------------------"
@@ -67,23 +85,22 @@ echo "--------------------------------------------------------------------------
 pip install dos2unix
 
 for file in $(find . -name "*.sh"); do
-    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.7/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
+    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.8/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
     mv ${file}_tmp ${file}
 done
 for file in $(find . -name "*.py"); do
-    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.7/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
+    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.8/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
     mv ${file}_tmp ${file}
 done
 for file in $(find . -name "*.c*"); do
-    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.7/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
+    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.8/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
     mv ${file}_tmp ${file}
 done
 for file in $(find . -name "*.h"); do
-    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.7/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
+    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.8/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
     mv ${file}_tmp ${file}
 done
 for file in $(find . -name ".fuse_hidden*"); do
-    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.7/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
+    python /opt/vitis_ai/conda/envs/vitis-ai-tensorflow2/lib/python3.8/site-packages/dos2unix.py ${file} ${file}_tmp > /dev/null
     mv ${file}_tmp ${file}
 done
-'
